+ add selectable Bearing info to Compass + Rework MapScale       -	don’t regenerate Texture       -	change width also with float values + refresh translations + save MapSettings with onHide + reactivate AccuracyCircle with setting option + fix NPE on SpoilerView with no selected Cache + fix NPE on GrabImagesSelectedByCache() with empty result from GC (like empty API key)
git-svn-id: svn://svn.code.sf.net/p/droidcachebox/code/trunk@1334 5890f383-b09f-485f-8f59-22bc967d67fd

+++ Cachebox Core/src/CB_Core/Map/Descriptor.java
@@ -167,7 +167,7 @@ public class Descriptor implements Comparable<Descriptor>
 	 *            Longitude
 	 * @return double
 	 */
-	public static double LongitudeToTileX(int zoom, double longitude)
+	public static double LongitudeToTileX(double zoom, double longitude)
 	{
 		return (longitude + 180.0) / 360.0 * Math.pow(2, zoom);
 	}
@@ -181,7 +181,7 @@ public class Descriptor implements Comparable<Descriptor>
 	 *            Latitude
 	 * @return double
 	 */
-	public static double LatitudeToTileY(int zoom, double latitude)
+	public static double LatitudeToTileY(double zoom, double latitude)
 	{
 		double latRad = latitude * Math.PI / 180.0;
 
@@ -191,7 +191,7 @@ public class Descriptor implements Comparable<Descriptor>
 	/**
 	 * Berechnet aus der übergebenen OSM-X-Koordinate den entsprechenden Längengrad
 	 */
-	public static double TileXToLongitude(int zoom, double x)
+	public static double TileXToLongitude(double zoom, double x)
 	{
 
 		return -180.0 + (360.0 * x) / Math.pow(2, zoom);
@@ -200,7 +200,7 @@ public class Descriptor implements Comparable<Descriptor>
 	/**
 	 * Berechnet aus der übergebenen OSM-Y-Koordinate den entsprechenden Breitengrad
 	 */
-	public static double TileYToLatitude(int zoom, double y)
+	public static double TileYToLatitude(double zoom, double y)
 	{
 		double xNom = Math.exp(2 * Math.PI) - Math.exp(4 * Math.PI * Math.pow(2, -zoom) * y);
 		double xDen = Math.exp(2 * Math.PI) + Math.exp(4 * Math.PI * Math.pow(2, -zoom) * y);
